Add skill dependency and finalize skill-builder component and service

+++ src/app/skill-builder/skill-unit/skill-unit.component.ts
@@ -9,7 +9,8 @@ import { SkillBuilderService } from '../skill-builder.service';
 })
 export class SkillUnitComponent implements OnInit {
   @Input() skill: Skill = {} as Skill;
-  @Input() isAvailable: boolean = false;
+  @Input() increasable: boolean = false;
+  @Input() decreasable: boolean = false;
   displayDescription = false;
 
   constructor(private skillBuilderService: SkillBuilderService) {}
@@ -17,28 +18,32 @@ export class SkillUnitComponent implements OnInit {
   ngOnInit(): void {}
 
   increaseSkillLevel() {
-    this.skillBuilderService.increaseUsedSkillPoints(
-      this.skill.skillPoints[this.skill.level]
-    );
-    this.skillBuilderService.increaseUsedHeroicPoints(
-      this.skill.heroicPoints[this.skill.level]
-    );
-    this.skillBuilderService.addMinLevel(
-      this.skill.requiredLevels[this.skill.level]
-    );
-    this.skill.level++;
+    if (this.increasable) {
+      this.skillBuilderService.increaseUsedSkillPoints(
+        this.skill.skillPoints[this.skill.level]
+      );
+      this.skillBuilderService.increaseUsedHeroicPoints(
+        this.skill.heroicPoints[this.skill.level]
+      );
+      this.skillBuilderService.addMinLevel(
+        this.skill.requiredLevels[this.skill.level]
+      );
+      this.skill.level++;
+    }
   }
 
   decreaseSkillLevel() {
-    this.skill.level--;
-    this.skillBuilderService.decreaseUsedSkillPoints(
-      this.skill.skillPoints[this.skill.level]
-    );
-    this.skillBuilderService.decreaseUsedHeroicPoints(
-      this.skill.heroicPoints[this.skill.level]
-    );
-    this.skillBuilderService.removeMinLevel(
-      this.skill.requiredLevels[this.skill.level]
-    );
+    if (this.decreasable) {
+      this.skill.level--;
+      this.skillBuilderService.decreaseUsedSkillPoints(
+        this.skill.skillPoints[this.skill.level]
+      );
+      this.skillBuilderService.decreaseUsedHeroicPoints(
+        this.skill.heroicPoints[this.skill.level]
+      );
+      this.skillBuilderService.removeMinLevel(
+        this.skill.requiredLevels[this.skill.level]
+      );
+    }
   }
 }
